+++ Lilu/Headers/kern_iokit.hpp
@@ -139,8 +139,9 @@ namespace WIOKit {
 		enum : uint16_t {
 			ATIAMD = 0x1002,
 			AMDZEN = 0x1022,
-			NVIDIA = 0x10de,
-			Intel = 0x8086
+			NVIDIA = 0x10DE,
+			Intel  = 0x8086,
+			VMware = 0x15AD
 		};
 	};