+++ Lilu/Sources/kern_devinfo.cpp
@@ -231,7 +231,8 @@ void DeviceInfo::grabDevicesFromPciRoot(IORegistryEntry *pciRoot) {
 			// Strip interface, as we only care about class and subclass
 			code &= WIOKit::ClassCode::PCISubclassMask;
 
-			if (!gotVendor || !gotClass || (vendor != WIOKit::VendorID::Intel && vendor != WIOKit::VendorID::ATIAMD && vendor != WIOKit::VendorID::AMDZEN))
+			if (!gotVendor || !gotClass || (vendor != WIOKit::VendorID::Intel && vendor != WIOKit::VendorID::ATIAMD &&
+			                                vendor != WIOKit::VendorID::AMDZEN && vendor != WIOKit::VendorID::VMware))
 				continue;
 
 			if (vendor == WIOKit::VendorID::Intel && (code == WIOKit::ClassCode::DisplayController || code == WIOKit::ClassCode::VGAController)) {
@@ -272,24 +273,27 @@ void DeviceInfo::grabDevicesFromPciRoot(IORegistryEntry *pciRoot) {
 									   safeString(pciobj->getName()), safeString(name),  pcivendor);
 								v.video = pciobj;
 								v.vendor = pcivendor;
-							} else if (pcicode == WIOKit::ClassCode::HDADevice && pcivendor != WIOKit::VendorID::AMDZEN) {
-								DBGLOG("dev", "found HDAU device %s at %s by %04X",
+							} else if (pcicode == WIOKit::ClassCode::HDADevice) {
+								DBGLOG("dev", "found audio device %s at %s by %04X",
 									   safeString(pciobj->getName()), safeString(name), pcivendor);
 								v.audio = pciobj;
-							} else if (pcicode == WIOKit::ClassCode::HDADevice) {
-								// On modern AMD platforms built-in audio devices sits on a PCI bridge just any other device.
-								// Luckily it has a distinct device-id for the time being.
-								DBGLOG("dev", "found AMD HDEF device %s at %s by %04X",
-									   safeString(pciobj->getName()), safeString(name), pcivendor);
-								audioBuiltinAnalog = pciobj;
 							}
 						}
 					}
 
 					pciiterator->release();
 
-					if (v.video && !videoExternal.push_back(v))
-						SYSLOG("dev", "failed to push video gpu");
+					if (v.video) {
+						DBGLOG_COND(v.audio, "dev", "marking audio device as HDAU at %s", safeString(v.audio->getName()));
+						if (!videoExternal.push_back(v))
+							SYSLOG("dev", "failed to push video gpu");
+					} else if (v.audio && !audioBuiltinAnalog) {
+						// On modern AMD platforms or VMware built-in audio devices sits on a PCI bridge just any other device.
+						// On AMD it has a distinct Ryzen device-id for the time being, yet on VMware it is just Intel.
+						// To distinguish the devices we use audio card presence as a marker.
+						DBGLOG("dev", "marking audio device as HDEF at %s", safeString(v.audio->getName()));
+						audioBuiltinAnalog = v.audio;
+					}
 				}
 			}
 		}